+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 1.1.1 (Next)
 - Improve WinRM password determination.
+- Ignore empty SSH private key in SSH backend arguments.
 
 ### 1.1.0
 - Support multiple TestInfra `test_files`.