+++ CHANGELOG.md
@@ -12,7 +12,7 @@
 ### 1.0.1
 - Testinfra WinRM backend support (beta).
 - Testinfra SSH backend improved support.
-- Prefer generated Host and Port over generated httpAddr
+- Prefer generated Host and Port over generated httpAddr.
 
 ### 1.0.0
 - Initial release.